Finish task of modules
